# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Rest API documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/logs": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Get all admins for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "activity",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
...
        "/api/etudiant": {
            "get": {
                "tags": [
                    "etudiant"
                ],
                "summary": "Get all etudiants for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "etudiantindex",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etudiant/create": {
            "post": {
                "tags": [
                    "etudiant"
                ],
                "summary": "create all etudiants for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "etudiantstore",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "description": "Book data that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_etudiant": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "prenom_etudiant": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "CNE": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "codeApogee": {
                                        "type": "integer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etudiant/edit/{etudiant}": {
            "put": {
                "tags": [
                    "etudiant"
                ],
                "summary": "update all etudiants for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "etudiantupdate",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "description": "codeApogee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Book data that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codeApogee": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "nom_etudiant": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "prenom_etudiant": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "CNE": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "id_rapport": {
                                        "type": "integer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/etudiant/{etudiant}": {
            "delete": {
                "tags": [
                    "etudiant"
                ],
                "summary": "delete all etudiants for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "etudiantdelete",
                "parameters": [
                    {
                        "name": "etudiant",
                        "in": "path",
                        "description": "codeApogee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Input your Bearer token in the following format - Bearer {your_token_here}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "securityDefinitions": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        },
        "security": [
            {
                "Bearer": []
            }
        ]
    }
}